adding controller
