+++ routes/admin.js
@@ -2,24 +2,12 @@ const express = require('express');
 
 const router = express.Router();
 
-const products = [];
+const productController = require('../controllers/products');
 
 // /admin/add-product => GET
-router.get('/add-product', (req, res, next) => {
-  res.render('add-product', {
-    pageTitle: 'Add Product',
-    path: '/admin/add-product',
-    formsCSS: true,
-    productCSS: true,
-    activeAddProduct: true
-  });
-});
+router.get('/add-product', productController.getAddProduct);
 
 // /admin/add-product => POST
-router.post('/add-product', (req, res, next) => {
-  products.push({ title: req.body.title });
-  res.redirect('/');
-});
+router.post('/add-product', productController.postAddProduct);
 
-exports.routes = router;
-exports.products = products;
+module.exports = router;
