+++ routes/shop.js
@@ -1,19 +1,9 @@
 const express = require('express');
 
-const adminData = require('./admin');
-
 const router = express.Router();
 
-router.get('/', (req, res, next) => {
-  const products = adminData.products;
-  res.render('shop', {
-    prods: products,
-    pageTitle: 'Shop',
-    path: '/',
-    hasProducts: products.length > 0,
-    activeShop: true,
-    productCSS: true
-  });
-});
+const productController = require('../controllers/products');
+
+router.get('/', productController.getProducts);
 
 module.exports = router;
